+++ plugins/main/public/components/index.js
@@ -21,7 +21,7 @@ import { KibanaVisWrapper } from '../components/management/cluster/cluster-visua
 import { ToastNotificationsModal } from '../components/notifications/modal';
 import { getAngularModule } from '../kibana-services';
 import { WzUpdatesNotification } from './wz-updates-notification';
-import { Cluster } from './management/cluster/cluster';
+import { GeneralCluster } from './management/cluster/general-cluster';
 
 const app = getAngularModule();
 
@@ -35,4 +35,4 @@ app.value('ClusterTimelions', ClusterTimelions);
 app.value('KibanaVisualization', KibanaVisWrapper);
 app.value('ToastNotificationsModal', ToastNotificationsModal);
 app.value('WzUpdatesNotification', WzUpdatesNotification);
-app.value('Cluster', Cluster);
+app.value('Cluster', GeneralCluster);
